refactor(desktop): modularize agent studio task tabs hook
split task tabs logic into persistence, selection, and actions hooks

add regression tests for hydration fallback and repo switch flows
