+++ apps/desktop/src/pages/agents/use-agent-studio-task-tabs-persistence.ts
@@ -1,0 +1,109 @@
+import type { TaskCard } from "@openducktor/contracts";
+import { type Dispatch, type SetStateAction, useEffect } from "react";
+import {
+  canPersistTaskTabs,
+  parsePersistedTaskTabs,
+  toPersistedTaskTabs,
+} from "./agents-page-session-tabs";
+import { toTabsStorageKey } from "./agents-page-utils";
+
+type SetState<T> = Dispatch<SetStateAction<T>>;
+
+type UseTaskTabPersistenceArgs = {
+  activeRepo: string | null;
+  taskId: string;
+  selectedTask: TaskCard | null;
+  tasks: TaskCard[];
+  isLoadingTasks: boolean;
+  openTaskTabs: string[];
+  tabsStorageHydratedRepo: string | null;
+  activeTaskTabId: string;
+  setOpenTaskTabs: SetState<string[]>;
+  setPersistedActiveTaskId: SetState<string | null>;
+  setIntentActiveTaskId: SetState<string | null>;
+  setTabsStorageHydratedRepo: SetState<string | null>;
+};
+
+export function useTaskTabPersistence(args: UseTaskTabPersistenceArgs): void {
+  const {
+    activeRepo,
+    taskId,
+    selectedTask,
+    tasks,
+    isLoadingTasks,
+    openTaskTabs,
+    tabsStorageHydratedRepo,
+    activeTaskTabId,
+    setOpenTaskTabs,
+    setPersistedActiveTaskId,
+    setIntentActiveTaskId,
+    setTabsStorageHydratedRepo,
+  } = args;
+
+  useEffect(() => {
+    if (!activeRepo) {
+      setOpenTaskTabs([]);
+      setPersistedActiveTaskId(null);
+      setIntentActiveTaskId(null);
+      setTabsStorageHydratedRepo(null);
+      return;
+    }
+
+    const raw = globalThis.localStorage.getItem(toTabsStorageKey(activeRepo));
+    const persistedTabs = parsePersistedTaskTabs(raw);
+    setOpenTaskTabs(persistedTabs.tabs);
+    setPersistedActiveTaskId(persistedTabs.activeTaskId);
+    setTabsStorageHydratedRepo(activeRepo);
+  }, [
+    activeRepo,
+    setIntentActiveTaskId,
+    setOpenTaskTabs,
+    setPersistedActiveTaskId,
+    setTabsStorageHydratedRepo,
+  ]);
+
+  useEffect(() => {
+    if (isLoadingTasks) {
+      return;
+    }
+    const knownTaskIds = new Set(tasks.map((task) => task.id));
+    setOpenTaskTabs((current) => {
+      const filtered = current.filter((taskTabId) => knownTaskIds.has(taskTabId));
+      if (filtered.length === current.length) {
+        return current;
+      }
+      return filtered;
+    });
+  }, [isLoadingTasks, setOpenTaskTabs, tasks]);
+
+  useEffect(() => {
+    if (!taskId) {
+      return;
+    }
+    if (!selectedTask) {
+      return;
+    }
+    setOpenTaskTabs((current) => {
+      if (current.includes(taskId)) {
+        return current;
+      }
+      return [...current, taskId];
+    });
+  }, [selectedTask, setOpenTaskTabs, taskId]);
+
+  useEffect(() => {
+    if (!canPersistTaskTabs(activeRepo, tabsStorageHydratedRepo)) {
+      return;
+    }
+    if (!activeRepo) {
+      return;
+    }
+    globalThis.localStorage.setItem(
+      toTabsStorageKey(activeRepo),
+      toPersistedTaskTabs({
+        tabs: openTaskTabs,
+        activeTaskId: activeTaskTabId || null,
+      }),
+    );
+  }, [activeRepo, activeTaskTabId, openTaskTabs, tabsStorageHydratedRepo]);
+}
