+++ apps/desktop/src/pages/agents/use-agent-studio-task-tabs.test.tsx
@@ -254,4 +254,188 @@ describe("useAgentStudioTaskTabs", () => {
       });
     }
   });
+
+  test("routes fallback tab after hydration when URL task is empty", async () => {
+    const memoryStorage = createMemoryStorage();
+    const originalStorage = globalThis.localStorage;
+    Object.defineProperty(globalThis, "localStorage", {
+      configurable: true,
+      value: memoryStorage,
+    });
+
+    try {
+      memoryStorage.setItem(
+        toTabsStorageKey("/repo"),
+        toPersistedTaskTabs({
+          tabs: ["task-2"],
+          activeTaskId: "task-2",
+        }),
+      );
+
+      const latestSession = createSession("task-2", "session-2");
+      const updateCalls: Array<Record<string, string | undefined>> = [];
+
+      const harness = createHookHarness({
+        activeRepo: "/repo",
+        taskId: "",
+        selectedTask: null,
+        tasks: [createTask("task-1"), createTask("task-2")],
+        isLoadingTasks: false,
+        latestSessionByTaskId: new Map([["task-2", latestSession]]),
+        updateQuery: (updates) => {
+          updateCalls.push(updates);
+        },
+        clearComposerInput: () => {},
+      });
+
+      await harness.mount();
+      await harness.waitFor(() => updateCalls.length > 0);
+
+      expect(updateCalls).toHaveLength(1);
+      expect(updateCalls[0]).toEqual({
+        task: "task-2",
+        session: "session-2",
+        agent: "spec",
+        scenario: "spec_initial",
+        autostart: undefined,
+        start: undefined,
+      });
+
+      await harness.unmount();
+    } finally {
+      Object.defineProperty(globalThis, "localStorage", {
+        configurable: true,
+        value: originalStorage,
+      });
+    }
+  });
+
+  test("does not route stale fallback tab while switching repos", async () => {
+    const memoryStorage = createMemoryStorage();
+    const originalStorage = globalThis.localStorage;
+    Object.defineProperty(globalThis, "localStorage", {
+      configurable: true,
+      value: memoryStorage,
+    });
+
+    try {
+      memoryStorage.setItem(
+        toTabsStorageKey("/repo-b"),
+        toPersistedTaskTabs({
+          tabs: ["task-b"],
+          activeTaskId: "task-b",
+        }),
+      );
+
+      const updateCalls: Array<Record<string, string | undefined>> = [];
+      const harness = createHookHarness({
+        activeRepo: "/repo-a",
+        taskId: "task-a",
+        selectedTask: createTask("task-a"),
+        tasks: [createTask("task-a")],
+        isLoadingTasks: false,
+        latestSessionByTaskId: new Map(),
+        updateQuery: (updates) => {
+          updateCalls.push(updates);
+        },
+        clearComposerInput: () => {},
+      });
+
+      await harness.mount();
+      expect(updateCalls).toHaveLength(0);
+
+      const repoBSession = createSession("task-b", "session-b");
+      await harness.update({
+        activeRepo: "/repo-b",
+        taskId: "",
+        selectedTask: null,
+        tasks: [createTask("task-b")],
+        isLoadingTasks: false,
+        latestSessionByTaskId: new Map([["task-b", repoBSession]]),
+        updateQuery: (updates) => {
+          updateCalls.push(updates);
+        },
+        clearComposerInput: () => {},
+      });
+
+      await harness.waitFor(() => updateCalls.length > 0);
+
+      expect(updateCalls[0]).toEqual({
+        task: "task-b",
+        session: "session-b",
+        agent: "spec",
+        scenario: "spec_initial",
+        autostart: undefined,
+        start: undefined,
+      });
+      expect(updateCalls.some((call) => call.task === "task-a")).toBeFalse();
+
+      await harness.unmount();
+    } finally {
+      Object.defineProperty(globalThis, "localStorage", {
+        configurable: true,
+        value: originalStorage,
+      });
+    }
+  });
+
+  test("clearing active repo skips fallback routing and resets tabs", async () => {
+    const memoryStorage = createMemoryStorage();
+    const originalStorage = globalThis.localStorage;
+    Object.defineProperty(globalThis, "localStorage", {
+      configurable: true,
+      value: memoryStorage,
+    });
+
+    try {
+      memoryStorage.setItem(
+        toTabsStorageKey("/repo"),
+        toPersistedTaskTabs({
+          tabs: ["task-1"],
+          activeTaskId: "task-1",
+        }),
+      );
+
+      const updateCalls: Array<Record<string, string | undefined>> = [];
+      const harness = createHookHarness({
+        activeRepo: "/repo",
+        taskId: "",
+        selectedTask: null,
+        tasks: [createTask("task-1")],
+        isLoadingTasks: false,
+        latestSessionByTaskId: new Map(),
+        updateQuery: (updates) => {
+          updateCalls.push(updates);
+        },
+        clearComposerInput: () => {},
+      });
+
+      await harness.mount();
+      await harness.waitFor(() => updateCalls.length > 0);
+      updateCalls.length = 0;
+
+      await harness.update({
+        activeRepo: null,
+        taskId: "",
+        selectedTask: null,
+        tasks: [],
+        isLoadingTasks: false,
+        latestSessionByTaskId: new Map(),
+        updateQuery: (updates) => {
+          updateCalls.push(updates);
+        },
+        clearComposerInput: () => {},
+      });
+
+      expect(updateCalls).toHaveLength(0);
+      expect(harness.getLatest().tabTaskIds).toEqual([]);
+
+      await harness.unmount();
+    } finally {
+      Object.defineProperty(globalThis, "localStorage", {
+        configurable: true,
+        value: originalStorage,
+      });
+    }
+  });
 });
